Factor out new `h.views.api.index` module
